Updated to totallylazy-815 and funclate-84

+++ test/com/googlecode/yatspec/plugin/sequencediagram/ByNamingConventionMessageProducerTest.java
@@ -22,7 +22,7 @@ public class ByNamingConventionMessageProducerTest {
     public void convertsValuesWithNamingConventionToSequenceDiagramMessages() {
         CapturedInputAndOutputs inputAndOutputs = new CapturedInputAndOutputs().add("Kiss from Boy to Girl", new Object()).add("Slap from Girl to Boy", new Object());
         Sequence<SequenceDiagramMessage> messages = sequence(new ByNamingConventionMessageProducer().messages(inputAndOutputs));
-        assertThat(messages.size().intValue(), is(equalTo(2)));
+        assertThat(messages.size(), is(equalTo(2)));
         assertThat(messages.first(), is(equalTo(new SequenceDiagramMessage("Boy", "Girl", "Kiss", "Kiss_from_Boy_to_Girl"))));
         assertThat(messages.second(), is(equalTo(new SequenceDiagramMessage("Girl", "Boy", "Slap", "Slap_from_Girl_to_Boy"))));
     }
@@ -31,7 +31,7 @@ public class ByNamingConventionMessageProducerTest {
     public void dealsWithGroups() {
         CapturedInputAndOutputs inputAndOutputs = new CapturedInputAndOutputs().add("(grouped) Kiss from Boy to Girl", new Object());
         Sequence<SequenceDiagramMessage> messages = sequence(new ByNamingConventionMessageProducer().messages(inputAndOutputs));
-        assertThat(messages.size().intValue(), is(equalTo(1)));
+        assertThat(messages.size(), is(equalTo(1)));
         assertThat(messages.first(), is(equalTo(new SequenceDiagramMessage("Boy", "Girl", "(grouped) Kiss", "_grouped__Kiss_from_Boy_to_Girl"))));
     }
 }
